Update file_ops.py
There are a few issues with the code:

The code does not include error handling for file opening and reading, which could result in errors if the file does not exist or is not readable.

In the third block of code, the split method is used to split the line into words using a comma as the delimiter. However, if the file does not contain commas, this could result in incorrect output.

In the fifth block of code, the list of cars is created by appending row, which is a list created by splitting the line using a comma as a delimiter. However, if the file does not contain commas, this could result in incorrect output.

The last block of code writes the list of row-lists to the output file as a string using the str method. However, this could result in unexpected output if the list contains nested lists or other complex data types.

The code does not close the file after writing to it, which could cause issues if the file is used by other processes or if the program is run multiple times.

In the last block of code, the 'a' mode is used to open the output file. This will append to the existing contents of the file, which may not be desired. If the goal is to overwrite the file, the 'w' mode should be used instead.