Update HexToDec.py
The code has a few issues:

The __getDecDigit() function is unnecessarily defined as a private function. Since it is not a method of a class, it can simply be defined as a regular function.
The __getDecDigit() function can be simplified by using the index() method of a list to find the index of the given digit.
The hexToDec() function is not returning the decimal number, it is only printing it. It should be modified to return the value instead.
The loop in the hexToDec() function is iterating over the length of the hex number in reverse order. This is unnecessary and can be simplified to a regular forward loop.